backend/services/azure_vision_service.py
import base64
import io
from PIL import Image
import logging
from typing import Dict, Any

logger = logging.getLogger(__name__)

class AzureVisionService:
    """Azure Computer Vision service for image analysis"""
    
    def __init__(self):
        # This would normally use Azure Computer Vision API
        # For now, we'll use mock responses since the main AI processing
        # is handled by LangChain service
        pass
    
    async def analyze_image(self, image: Image.Image) -> Dict[str, Any]:
        """Analyze image using Azure Computer Vision (mocked for development)"""
        try:
            # Mock image analysis response
            # In production, this would call Azure Computer Vision API
            
            # Basic image validation
            width, height = image.size
            if width < 100 or height < 100:
                return {
                    "error": "Image too small for analysis",
                    "details": "Please upload a clearer, larger image"
                }
            
            # Mock analysis result
            mock_result = {
                "description": "Agricultural crop image detected",
                "objects": ["plant", "leaves", "crop"],
                "colors": ["green", "brown", "yellow"],
                "text": [],  # OCR results if any
                "faces": [],  # Should be empty for crop images
                "brands": [],  # Brand detection
                "image_quality": "good"
            }
            
            return mock_result
            
        except Exception as e:
            logger.error(f"Error in Azure Vision analysis: {str(e)}")
            return {
                "error": "Image analysis failed",
                "details": str(e)
            }
    
    def validate_crop_image(self, image: Image.Image) -> bool:
        """Validate if image is suitable for crop analysis"""
        try:
            # Basic validations
            width, height = image.size
            
            # Size check
            if width < 200 or height < 200:
                return False
            
            # File size check (approximate)
            buffer = io.BytesIO()
            image.save(buffer, format='JPEG')
            size_mb = len(buffer.getvalue()) / (1024 * 1024)
            
            if size_mb > 10:  # 10MB limit
                return False
            
            return True
            
        except Exception as e:
            logger.error(f"Error validating image: {str(e)}")
            return False
