backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.10
openai>=1.10.0,<2.0.0
azure-cognitiveservices-speech==1.34.0
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
pydantic==2.5.2
pydantic-settings>=2.4.0,<3.0.0
httpx==0.25.2
